# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/whatsapp-agents/me": {
      "get": {
        "operationId": "WhatsAppAgentController_getMyAgent",
        "summary": "Get current user's WhatsApp agent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the WhatsApp agent for the current user"
          },
          "404": {
            "description": "WhatsApp agent not found for this user"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/whatsapp-agents/{id}/status": {
      "patch": {
        "operationId": "WhatsAppAgentController_updateAgentStatus",
        "summary": "Update agent status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "WhatsApp agent ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent status successfully updated"
          },
          "404": {
            "description": "WhatsApp agent not found"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/whatsapp-agents/labels": {
      "get": {
        "operationId": "WhatsAppAgentController_getLabels",
        "summary": "Get all WhatsApp labels for current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all WhatsApp labels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "hexColor": {
                        "type": "string"
                      },
                      "colorIndex": {
                        "type": "number"
                      },
                      "count": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "WhatsApp agent not found"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/whatsapp-agents/contacts/validate": {
      "post": {
        "operationId": "WhatsAppAgentController_validateContact",
        "summary": "Validate if a phone number exists on WhatsApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns validation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "contactId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "WhatsApp agent not found"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/whatsapp-agents/config": {
      "patch": {
        "operationId": "WhatsAppAgentController_updateAgentConfig",
        "summary": "Update agent configuration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent configuration updated successfully"
          },
          "404": {
            "description": "WhatsApp agent not found"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent/can-process": {
      "post": {
        "operationId": "AgentController_canProcess",
        "summary": "Check if agent can process a message",
        "description": "Called by whatsapp-agent service to check if a message should be processed and get agent context + authorized groups",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanProcessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns processing decision and agent configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanProcessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "agent"
        ]
      }
    },
    "/agent/log-operation": {
      "post": {
        "operationId": "AgentController_logOperation",
        "summary": "Log an agent operation with full metrics",
        "description": "Called by whatsapp-agent service to log conversations with tokens, tools, and duration metrics",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogOperationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operation logged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "operationId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "agent"
        ]
      }
    },
    "/agent-internal/agents/me": {
      "get": {
        "operationId": "WhatsAppAgentInternalController_getAgentSnapshot",
        "summary": "Lire l'état interne complet de l'agent",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent pour récupérer l'objet WhatsAppAgent complet (cache local côté agent) et le groupe de gestion associé.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Snapshot complet de l'agent retourné"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "WhatsAppAgentInternalController_updateAgentSnapshot",
        "summary": "Mettre à jour l'état interne de l'agent",
        "description": "Endpoint interne backend unique pour les mises à jour venant du whatsapp-agent. Toutes les propriétés sont optionnelles (prompt, statut de sync image, erreur sync).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentInternalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Snapshot agent mis à jour et retourné"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "whatsapp-agents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateAgentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PROVISIONING",
              "RUNNING",
              "STOPPED",
              "ERROR",
              "DELETED"
            ],
            "description": "The new status of the agent",
            "example": "RUNNING"
          },
          "connectionStatus": {
            "type": "string",
            "enum": [
              "CONNECTED",
              "DISCONNECTED",
              "QR_REQUIRED",
              "PAIRING_REQUIRED"
            ],
            "description": "The connection status of WhatsApp",
            "example": "CONNECTED"
          }
        },
        "required": [
          "status"
        ]
      },
      "ValidateContactDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number to validate",
            "example": "+237612345678"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "UpdateAgentConfigDto": {
        "type": "object",
        "properties": {
          "testPhoneNumbers": {
            "description": "Phone numbers for test mode (AI only responds to these)",
            "example": [
              "+237612345678",
              "+237698765432"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "testLabels": {
            "description": "Label names for test mode (AI only responds to contacts with these labels)",
            "example": [
              "Test",
              "VIP"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "labelsToNotReply": {
            "description": "Label names to exclude in production mode (AI will not respond to contacts with these labels)",
            "example": [
              "DoNotReply",
              "Blocked"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productionEnabled": {
            "type": "boolean",
            "description": "Enable production mode for the AI agent",
            "example": true
          }
        }
      },
      "ContactLabelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Label ID"
          },
          "name": {
            "type": "string",
            "description": "Label name"
          },
          "hexColor": {
            "type": "string",
            "description": "Label hex color"
          }
        },
        "required": [
          "id",
          "name",
          "hexColor"
        ]
      },
      "CanProcessDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Connected WhatsApp account ID (owner of the WhatsApp instance)",
            "example": "237657888690@c.us"
          },
          "chatId": {
            "type": "string",
            "description": "WhatsApp chat ID (where the message was received)",
            "example": "237657888690@c.us"
          },
          "message": {
            "type": "string",
            "description": "User message",
            "example": "Bonjour, je voudrais commander un produit"
          },
          "contactLabels": {
            "description": "Labels of the contact sending the message",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactLabelDto"
            }
          },
          "timestamp": {
            "type": "string",
            "description": "Message timestamp",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "userId",
          "chatId",
          "message",
          "timestamp"
        ]
      },
      "AuthorizedGroupDto": {
        "type": "object",
        "properties": {
          "whatsappGroupId": {
            "type": "string",
            "description": "WhatsApp group ID",
            "example": "12345678@g.us"
          },
          "usage": {
            "type": "string",
            "description": "Usage description of the group",
            "example": "Support client"
          }
        },
        "required": [
          "whatsappGroupId",
          "usage"
        ]
      },
      "CanProcessResponseDto": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean",
            "description": "Whether the agent can process this message",
            "example": true
          },
          "reason": {
            "type": "string",
            "description": "Reason if not allowed",
            "example": "Agent not configured"
          },
          "agentContext": {
            "type": "string",
            "description": "Agent context (business instructions)",
            "example": "Vous êtes un assistant commercial..."
          },
          "managementGroupId": {
            "type": "string",
            "description": "Management group ID for forwarding",
            "example": "12345678@g.us"
          },
          "agentId": {
            "type": "string",
            "description": "Agent ID",
            "example": "clx123456"
          },
          "authorizedGroups": {
            "description": "List of authorized groups",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorizedGroupDto"
            }
          }
        },
        "required": [
          "allowed"
        ]
      },
      "ToolExecutionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tool name"
          },
          "args": {
            "type": "object",
            "description": "Tool arguments"
          },
          "result": {
            "type": "object",
            "description": "Tool result"
          },
          "error": {
            "type": "string",
            "description": "Error message if tool failed"
          },
          "durationMs": {
            "type": "number",
            "description": "Tool execution duration in ms"
          }
        },
        "required": [
          "name",
          "args"
        ]
      },
      "LogOperationDto": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "description": "WhatsApp chat ID",
            "example": "237657888690@c.us"
          },
          "agentId": {
            "type": "string",
            "description": "Agent ID",
            "example": "clx123456"
          },
          "userId": {
            "type": "string",
            "description": "User ID",
            "example": "clx789012"
          },
          "userMessage": {
            "type": "string",
            "description": "User message",
            "example": "Bonjour, je voudrais commander un produit"
          },
          "agentResponse": {
            "type": "string",
            "description": "Agent response",
            "example": "Bonjour ! Je vais vous aider avec votre commande."
          },
          "systemPrompt": {
            "type": "string",
            "description": "System prompt used",
            "example": "Vous êtes un assistant commercial..."
          },
          "totalTokens": {
            "type": "number",
            "description": "Total tokens used",
            "example": 150
          },
          "promptTokens": {
            "type": "number",
            "description": "Prompt tokens",
            "example": 100
          },
          "completionTokens": {
            "type": "number",
            "description": "Completion tokens",
            "example": 50
          },
          "durationMs": {
            "type": "number",
            "description": "Operation duration in milliseconds",
            "example": 2500
          },
          "modelName": {
            "type": "string",
            "description": "Model name used",
            "example": "grok-beta"
          },
          "toolsUsed": {
            "description": "Tools used during execution",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolExecutionDto"
            }
          },
          "status": {
            "type": "string",
            "description": "Operation status",
            "example": "success",
            "enum": [
              "success",
              "error",
              "rate_limited"
            ]
          },
          "error": {
            "type": "string",
            "description": "Error message if failed"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata"
          },
          "timestamp": {
            "type": "string",
            "description": "Operation timestamp",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "chatId",
          "userMessage",
          "agentResponse",
          "systemPrompt",
          "durationMs",
          "status",
          "timestamp"
        ]
      },
      "UpdateAgentInternalDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}